services/edgar-mcp/pyproject.toml
[project]
name = "edgar-mcp"
version = "0.1.0"
description = "Edgar MCP service with FastMCP"
requires-python = ">=3.12"
dependencies = [
    "fastmcp~=3.0.0b1",
    "fastapi~=0.128.0",
    "uvicorn[standard]~=0.40.0",
    "pydantic~=2.12.5",
    "pydantic-settings~=2.12.0",
    "python-dotenv~=1.2.1",
    "sqlalchemy[asyncio]~=2.0.46",
    "asyncpg~=0.31.0",
    "alembic~=1.18.3",
    "aio-pika~=9.5.8",
    "cloudevents[pydantic]~=1.12.0",
    "structlog~=25.5.0",
    "opentelemetry-api~=1.39.1",
    "opentelemetry-sdk~=1.39.1",
    "opentelemetry-exporter-prometheus~=0.60b1",
    "opentelemetry-instrumentation-fastapi~=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy~=0.60b1",
    "opentelemetry-instrumentation-aio-pika~=0.60b1",
]

[project.optional-dependencies]
dev = [
    "basedpyright~=1.37.4",
    "pytest~=9.0.2",
    "pytest-asyncio~=1.3.0",
    "httpx~=0.28.1",
    "testcontainers[postgres,rabbitmq]~=4.14.1",
    "ruff~=0.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning:testcontainers"]

[tool.hatch.build.targets.wheel]
packages = ["src/edgar_mcp"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN2"]

[tool.basedpyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
extraPaths = ["src"]
typeCheckingMode = "all"
reportUnusedCallResult = false
reportMissingTypeStubs = false
